requirements-venv.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OCR y procesamiento de imágenes
pytesseract==0.3.10
pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3

# Cloud OCR APIs (opcionales)
google-cloud-vision==3.4.4
boto3==1.34.0

# Procesamiento de lenguaje natural
spacy==3.7.2
regex==2023.10.3
openai==1.3.0
langchain==0.0.340
transformers==4.35.2

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache y procesamiento asíncrono
redis==5.0.1
aioredis==2.0.1
rq==1.15.1

# Utilidades
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
pandas==2.1.4

# Desarrollo y testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Dependencias adicionales para entorno virtual
wheel==0.42.0
setuptools==69.0.2
